initial setup of project
Added details to the README file about this project, along with
additional documents files for tools and windows (not yet supported) in
the docs directory.

Added an initial main CMakeLists.txt file, which checks the CMake
version, turns on all warnings and errors, checks that the compiler
works for C++11 (basic support in test/cpp11.cpp), verifies that
building is not occurring in the source directory (a build sub-directory
is acceptable), checks the build type and echos it (allows an
environment variable to be set), checks for coverage enabling (verifies
Debug build type, and sets compiler options and coverage library),
enables testing and added the test sub-directory.

The test sub-directory contains the single Catch header file along with
a catch_main.cpp source file (prevents long compiling of the Catch
header file for main() function generation as it only needs to be
compiled once), and a parse number test source file (which has a single
test for a proof of concept that the test will be successfully built and
will run).