const-num: added initial test for empty token when no number
TEST: return an empty token when no number is present in an input
stream.  Enough was implemented to make this code compile, run and pass.

Added a new constant number parser class that will parse numerical
constants from an input stream.  It will return a unique pointer to a
token containing the the number constant if one is present or an a empty
token (null pointer) if the input stream does not contain a number.
(No white space is skipped as this will occur higher up.)

The initial test passes a non-number ("A") to the parser and expects an
empty token.  A minimum token class was implemented, but is nothing more
than a constructor and a virtual destructor (because it is going to be a
base class of different token classes).

The constant number parser class has an input stream member set in the
constructor, but is not yet used.  I was added so that the "A" could be
passed to the parser.  The get token member simply returns an empty
token (null pointer).  This was enough to make the initial test pass.

Other: The adding of the test was moved to the main cmake file so that
the source files can easily be reached.  For now all the source and
header files will be put in the root directory (a directory tree will be
created as the number of files increases).  The first parser number unit
test source file was removed, replaced with a new constant number unit
test source file.  Catch was updated to version 1.5.4.