const-num: added reading of a single digit to get non-empty token
TEST: return a token when a single digit is present.  Enough code was
implemented to return a token, though technically empty, the unique
pointer has a non-null pointer in it - nothing is actually returned in
the token yet.

The state pattern is being used in the constant number parser, which
contains a pointer to the current state instance.  The base state class
is abstract and contains virtual functions to process a character (with
a reference to the parser and the character to process as arguments) and
a done to report when input is done.  Each state class will contain a
static function for returning an instance to the class (using the
singleton pattern where the single instance is a state local variable of
the instance function).  The state classes contain no members.

The constant number parser was also given a string member to hold the
characters of the number parsed.  A change state function allows the
state to be changed and an add function to add a character to the number
string.  The state is initialized to the start state instance.

Two states were implemented, at start and an end state class.  The start
class just checks if the character is a digit and if it is, calls a new
add function in the parser to add the digit to the number string and
changes the state to the end state instance, otherwise if does nothing.
Both states return true for the done function (if the state is still the
start state, then the input did not contain a valid number character).

A process input function was added to the parser class with a loop that
gets characters from the input stream, calls the current state's process
function to process the character, and continues the loop if the state
does not report done.  After calling this function, the get token
returns an empty token if the number string is empty, other wise creates
a token to return.

OTHER: added a missing include of iosfwd (missing but appeared to work
anyway.