const-num: refactored number parsing done mechanism
It was noticed by looking at the coverage of the code so far that the
process function of the end state class was not being called.  Which was
expected since the purpose of the end state is to return true to the
done function, and its process function does nothing anyway.

Looking forward, all of the other states (yet to be added) will have
their done function return false.  This will be a bunch of functions
that are not really necessary.

Therefore, the done virtual function was removed from the state class
hierarchy.  In it place is a new done boolean member to the constant
number class along with an access function to set it (this flag is
initialized to false in the constructor).  Now if a state wants to
terminate parsing, it calls the set done function.

The loop in the process input function was changed to check this flag
instead of calling the state done function.  The start state process
function was changed to call the set done function.  The end state class
was removed since it is no longer necessary and the change state call
in the start state process function was removed.

The change state function is currently not called, but it will be needed
so it was not removed.

OTHER: the initialization of the constant number parser member variables
was changed to the C++11 initializer syntax.