const-num: started changed from token model to code model
It was realized that the token model used in the IBCP is not necessary
for the new model being used for IBCX.  The BASIC program source code
can be directly converted (parsed, translated, and encoded) into
internal program code.

Expressions will be converted using the recursive method instead of the
prioritized table method used for IBCP, which require a lot of complex
logic to accomplish the conversion.

In the first step in removing the token code implemented so far, the
first test (for a null token) was modified to call a new get code
function, which will take a reference to a program code instance.  The
get code function will return a data type.  A null data type will
indicate that nothing was parsed.  The minimum code necessary to make
this compile and run successfully was implemented.

The minimum data type class enumeration was defined (just a null
enumerator).  It was put into its own header file.  Unlike with IBCP,
more file granularity will be used with IBCX - files will only need to
include what headers they need without baggage they do not need.  This
should help with compile times.

A minimum program code class was implemented, which just consists of a
default constructor (does nothing) and a vector of program words.  A
minimum program word class was implemented, which consists of a default
constructor (does nothing).

The constant number parser get code function just returns a null data
type, which was all that was necessary to make the first test pass, plus
have the interface that will be needed (otherwise it would not have
needed the program code argument).