const-num: changed second integer test to get code model
As per the TDD process, only the minimum changes were made to make the
modified test pass, which was a change to the get code function to
return a null data type if the number string is empty, or an integer
data type if the number string is not empty.

This change allowed the removal of the get token function along with the
token source files.