const-num: started generating code words
TEST: verify the number of code words, none if not a number, two if has
a number.

The minimum code to make these tests pass was to just add two dummy
words (zeroes) to the program code vector.  This required access
functions to the program code class to determine if the program code
vector is empty, return the size of the vector and to append (emplace
back) words to the vector.  The default program word constructor was
replaced with one taking a word argument, which is used to initialize
the new word member that was added.