const-num: added actual code value into program code vector
TEST: verify the correct code value is put into the first word of the
constant integer program code.

Implemented a new code class, which assigned codes automatically as each
instance is initialized (the code class contains a static count variable
to maintain a unique code value for each code instance).  An instance of
this class was added to the constant number parser source file for a
constant integer.

A program word constructor was added taking a code argument, which gets
the value from the code to put into the program word.  An instruction
code access function was added to get the code.  A [] operator function
was added to the program code class to return code words.