const-num: added actual operand value into program code vector
TEST: verify that the original number string is stored into the second
word of the constant integer program code.

This required the number string to be stored in a dictionary where the
index within the dictionary is stored in the program code.

A minimum dictionary class was implemented containing a key map where
the key is the string (number in this case) to be stored and the data
value is the index of the entry.  The iterator to the entry is stored in
an iterator vector within the dictionary (iterators of maps can be
stored since the iterators don't change as entries are added to the
map).  The dictionary class contains a basic add member function that
returns the index of the entry, and a get function that returns the
string stored in the dictionary for an index.

A minimum program mode class was implemented for containing the constant
integer dictionary, and provides and access function to it.  The
constant number parser get code function was modified to take an
additional reference to the program model, which is used to add the
number string to the dictionary and get its index to put into the second
word of the constant integer code.

An operand access function was added to the program word.  Constant
integer codes are now fully encoded into the program vector.