const-num: added support for multiple digit integer constants
TEST: verify a string with multiple digits are parsed correctly.

Implemented a new mantissa state class in the constant number parser
source file.  The start state process function was modified to change
the state to the mantissa state for a digit, and now only calls the set
done function if a non-digit is processed.  The mantissa state process
function adds digits and calls the set done function otherwise.

OTHER: The state instance functions were changed to return a pointer to
a state instead of its own class pointer.