const-num: properly parse and of number
TEST: the next character in the input stream after reading the number is
correct.

The loop in the process input function was implemented where is takes
the next character from the input stream, and processes it.  However, if
the character is not actually consumed for the number string, it is
lost.

The new test for "345+" where the next character in the input stream
should have been "+" initially failed because a -1 was returned for the
next character (in other words, end of stream).

The process input function was modified to only peek at the next
character in the input stream and process that.  The add function, which
was more appropriately renamed add next character, was modified to get
the next character in the input stream and add it to the number string.
The character argument was no longer needed, and was removed.

OTHER: the 'c' variable in the process input function and argument of
the state process functions was renamed to 'next character' to better
reflect what it is (single letter 'c' was a poor choice to begin with,
which is too short).