const-num: parse a number with a decimal point
TEST: parse a number with a decimal point and verify data type returned
is a double.

The minimum changes to make this test pass include changing the mantissa
state process function to check for a '.' decimal point character and
set a flag in the parser to indicate the number is floating point along
with adding the character to the number string.  A set double access
function was added to the parser.

In the get code function, if this new flag is set, then the new double
data type enumerator is return, otherwise the integer data type is
returned.

Note: at the moment, the number is still [incorrectly] put into the
integer constant dictionary, but that was not part of this test.

OTHER: change the string for the integer test case for consistency.