const-num: parse a number with a decimal point at the beginning
TEST: parse a number with a decimal point at the beginning and verify
data type returned is a double.

Previously, a decimal point was only valid after a digit or negative
sign was seen (in other words, only checked in the mantissa state).  To
make this test pass, the start state also needs to check for a decimal
point and call the set double function.

OTHER: the code was refactor (also mantissa state process function) to
not duplicate the call to add next character (and in the case of start
state process, the call to the change state function).