const-num: stop parsing at a second decimal point
TEST: verify that when a second decimal point appears that parsing stops
at the second decimal point, which is the next character.

Previously, additional decimal points were allowed as it the mantissa
state process function simply was accepting periods as valid.  To make
this test pass, a check was added to only accept a period if the number
if not already marked as a double, which required a new is double access
function.

Note: a second period is technically an error, but this will either by
caught upstream when the correct end of the expression is checked or by
the number parsing error handling if two periods appear before any digit
characters.