const-num: parse a number with an exponent
TEST: verify that a double is returned that contains an 'e' character.

All that was needed to make this test pass was to modify the mantissa
state process function to check if the next character is an 'e' after
converting it to upper case (to allow 'e' and 'E'), and set the double
flag.

Note: this change is not sufficient to validate that the number is in
the correct format (mantissa, 'e', optional +/- sign, exponent digits),
but all that is needed is to make the test pass, which the change above
accomplishes.