const-num: accept only one 'e' for an exponent
TEST: verify that a second 'E' terminates parsing where the 'E' should
be the next character in the stream.

Previously, all 'e' characters would be accepted as part of the number,
and would have been consumed by the parser into the number string.

A new exponent state was added where its process function only accepts
digits, otherwise it calls the set done function to terminate the
parsing.