const-num: stop parsing at a second minus character
TEST: verify that when a second minus character terminates the parser
and will be the next character

Previously, a minus character was accepted anywhere in the exponent.
This test was made to pass by adding a new exponent digits state that is
changed to once a digit or minus character is processed.  The process
function of the exponent state only accepts digit characters.

As revealed by the coverage tests, the set done function is no longer
called in the exponent state process function.  Any character other
should cause an error (except a plus character, which hasn't been added
yet).  Errors have also not been implemented yet, so this call was
removed as it needs to be replaced.

OTHER: the names of the static state instance for each of the state
classes were renamed to simply 'state'.  This will make it easier to
clone this code for additional states.