const-num: throw exception if constant starts with unknown character
TEST: pass a string that started with a non-constant character and
verify that a parse error exception is thrown.

The original first test with string "A" that tested for a null token
pointer, which was modified to check for a null data type (and no
program code generated) needed to be removed is this was not a valid
test (and the description string was wrong anyway).  Eventually,
operands that start with letters (variables, array, function names,
etc.) will be process before number constants get parsed.

The new test was put into a new test case that will be testing for all
types of constant errors, and only checks that the error was thrown.

To make this test pass (and first compile), a simple empty parse error
structure was added (new source and header file).  The start state
process function was modified to throw this structure instead of setting
the done flag.

With this change, the get code function can no longer return a null data
type, so the check for an empty number string (which triggered this
return) was removed.  This was discovered by checking code coverage.