program: renamed program model class to program unit
The name "program unit" was the name of the class in the IBCP base on
a Qt abstract model (but this should have been named "program model").
It is used as a program unit (the idea is that each subroutine and user
function would be contained in a separate program unit).  The name
"program model" will be reserved for that purpose.

The name "program unit" will therefore be a separate non-Qt class for
holding the main routine, sub-routines and user functions, and the class
name and files were renaming accordingly.  The program unit and code
arguments to the get code function was also swapped as this is more
logically - the program unit will contain program code, and the program
code will hold a line of code (the code argument was renamed to code
line to more accurately reflect what it is).