project: removed trailing '_' character from constructor arguments
It turns out that there is no need to differentiate the argument name of
a constructor with the member name that it is initializing (the trailing
'_' was used to accomplish this).  With the IBCP, member variables were
prefixed with 'm_' to distinguish member variables, so this alternate
solution was not needed.  However, this is archaic and is not being used
in this repository.  Modern IDEs can distinguish arguments from member
variables by color.  And with the compiler warning options being used,
any mistakes will generate a compiler warning.

This was also used in the change state function argument.  This argument
was renamed new state (using the 'this->' method is ugly and won't be
used).