const-num: return null data type for non-number instead of throwing
TEST: verify that parser returns null data type when no number is
present in the input stream.

This is returning to the way it previously work versus the recent change
making it return an error.  The other error test was temporarily removed
since no error is currently generated (the parser error class remains as
it will be used).

This decision was made so that the function calling the number parser
can decide how to respond when no number is present.  The error
"expected numerical constant or unary operator" suggests that it should
be the expression parser's (not yet implemented) responsibility to throw
this particular error.  It is conceivable that some future command may
what to respond differently if no number is present.

The originally functionality was put back in where the start process
sets the done flag, and the get code function returns a null data type
if the number string is empty.