project: changed name of project slightly
Changed from 'Interactive' to 'Incremental' BASIC compiler to better
reflect what this project is trying to accomplish.  And its an
'Experiment' instead of '- Experimental'.

This project will become a library.  If used with a GUI (for example,
replacing the current core code in the IBCP), it will be the GUI that
will make it 'Interactive'.  As an interim, it is conceivable that this
library could also be used with a command line utility that simply
compiles the code and optionally runs it.  No GUI would be involved so
'Interactive' does not apply.  And 'Incremental' (what this type of
compiler is after all) happens to have the same first letter.