const-num: insure leading zero digit is followed by a period
TEST: verify the another digit following a loaded zero digit throws an
error.

To make this test pass, a new zero state class was added, which the
start state sets for a zero '0' character.  The zero state process
function checks if the next character is a period, which cause the
double flag to be set, the state changed to mantissa and the character
to be added to the number string.  If the next character is a digit, a
parser error thrown.

The structure of start state process function was rearranged a bit to
eliminate some possible code duplication (changing the state to the
mantissa state).

OTHER: removed the done function declaration from the start state class
that was missed; and temporary removed the contents of the parse error
class which is not being used yet (necessary to make the code compile)
along with its source file (which may not be needed).