const-num: added a column member to the parse error class
TEST: check for the column of the error (the digit after the leading
zero) for the leading zero followed by a digit error.

A column member was added to the parse error class and a column argument
added to the constructor.  To make the test case, a '1' argument was
added to the throw.  (This is not correct, but it does make the test
pass, which all that is needed.)