const-num: throw correct column for parse error
TEST: change input string to have something in front of the number,
which is skipped over because calling the number parser to make sure the
correct column is returned when the constant does not start at the
beginning of the input stream.

The simple '1' for the column does not work when the constant isn't at
the beginning of the input stream.  To make this test pass, the zero
state process function needs the column of the current character in the
input stream.  This can be obtained from the input stream, but the zero
state process function does not have access to the input stream.  A
get column access function was added to the parser class to provide this
access.

OTHER: added a blank line after the get code separating the rest of the
functions that are intended to be used by the state classes.