const-num: end when leading zero followed by non-constant character
TEST: verify that parsing ends when a non-period and non-digit character
follows a leading zero.

The zero state process function needed to set the done flag when the
next character was not a period or a digit.  This was not included
before because it was needed to make the previous tests pass.  A test
has now been added to check this.

OTHER: renamed the get code function to parse - get code seems to imply
a simple access function, and this function does more than that.  Also,
the arguments were reversed with code line now the first argument
(again) because it is the target of the parse (from the input stream);
the program object is for reference only.