const-num: cleaned up test descriptions
Moved test for a non-constant in the input stream back to the first test
where it was originally.  Decided the errors test case should the zero
test case since it will become rather long if it is going to include all
of the error tests, and changed the test case and section messages
accordingly.