const-num: check error if leading period not followed by a digit
TEST: verify that a non-digit after a leading period throws an error
(period selected as the second character since it can be part of a
constant, so this also verifies the second period is not incorrectly
accepted as valid).

To make this test pass, a new period state class was added.  If a period
is the first character, the start state process function changed to the
period state.  The period state process function checks for a digit and
changes to the mantissa state.  Otherwise if throws an error.

Note: curiously in the period state process function, the digit doesn't
need to be added to the number string by calling the add next character
function.  When changing to the mantissa state and leaving the character
in the input stream, when it gets to the mantissa state process
function, it will add the digit to the number string.