const-num: rearranged state functions, cleaned up unit tests
The zero and period state functions were put into the wrong spot with
respect to their class definitions and the order they are used.  The
state process functions were put together in the same order as the
class definitions.  The instance functions (the singleton pattern) was
overkill, so these were replaced with a set of static instances, one
for each state class.  The change state function was changed to take a
reference to a state, which a pointer is taken to put into the state
member pointer.  These changes make the code cleaner.

The unit tests contained a lot of duplication (checking the data type,
code size, getting the operand, and getting the dictionary entry for the
operand).  So this duplication was put into two new local macros, one
got integer operands and one for double operands.  Replacing the code
with these macros added a few tests that weren't present before.  Also
added a check for a zero code line size to the test for no constant
present.