const-num: allow for possibility of EQV operator not an exponent
TEST: verify that an operator that starts with 'E' (only EQV) does not
produce an error at the second latter.

In the exponent state process function, for a non-sign or non-digit, if
the next character is a letter, it may be the start of an operator
(currently only the EQV operator, but allow for any word operator), so
it should not through an error.  In addition, the 'E' would have already
been added to the number, therefore it needs to be erased from the
number string, which is accomplished by a new set possible operator
function, which also sets the done flag.

For a string like "1EQ", the 'E' sets the double flag, but in this case
the number "1" is not floating point.  To prevent this problem, the
double flag is now not set until a valid exponent character after the
'E' is found.

Note: the 'E' character is lost, this will be handled later - the parser
will contain a flag indicating this condition, and when it comes time to
look for an operator, if this flag is set, an 'E' will be assumed to
have been there before the reset of the operator's characters are
obtained from the input stream.