const-num: make sure digit follows exponent sign
TEST: verify that a digit follows the optional exponent sign.

To make this test pass, a new exponent sign state class and instance was
added where its process function checks for a digit changing to the
exponent digits state otherwise throwing a parse error (blank since that
is all that is needed to make the test pass, and the next test for the
error message initially fail).

The exponent state process function was changed where a '-' or '+' sign
character now changes to the exponent sign state, and a digit changing
to the exponent digits state.  After verifying the test passed, the
function was rearranged a bit to eliminate some statement duplication.