const-num: correct error column when at end of input stream
TEST: verify the error column is correct when at the end of the input
stream (the next character is EOF).

A problem was found with the last test when checking the error column
with the test string set to "1e-".  The column returned was -1
(0xFFFFFFFF since it was unsigned).  This occurred when trying to peek
at the next character and there is no next character.  The peek
character will be set to EOF, but this sets an error in the input
stream.  So when getting the current position in the input stream, the
-1 was returned, which is the function that the get column access
function was using.

To correct this problem, the current position of the next character
needed to be obtained before peeking at the next character in the input
stream.  A statement was added to the do-while loop in the process input
function before obtaining the next character to save the current
position into a new column member.  The get column access function was
changed to return the value of this column member.