const-num: look for possible negate operator
TEST: verify that no constant is returned if the input stream starts
with a minus '-' character that is not followed by a valid constant
character (digit).  This condition indicates the negate operator and not
a negative constant.  The "-e" was selected for this test since an 'E'
character is valid constant character, but only when preceded by a valid
mantissa (which is not the case with "-e").  Also checks that no code is
added.

To handle the character after the minus character, a new negative state
class and static instance was added where its process function checks
for a digit changing to the mantissa state otherwise terminating the
parser.  The start state process function sets this state when a minus
character is seen.

Terminating the parser required more then just setting the done flag
since the minus character had already been to the number string and with
the number string not empty, a number was returned.  A new set negate
function was added to first clear the number string and then set the
done flag.

Note: the constant number parser will eventually need to convey this
information (a negate operator was seen, and also the possibility the
the number is followed be an operator starting with 'E') back to the
caller.