const-num: allow a period after a minus sign
TEST: verify that period after a minus character is seen as a valid
constant character.

When code coverage was run on the previous change, it was seen that the
code for checking a period in the negative state process function was
not executed.  Since this code was not required to make the previous
test pass, it was removed (temporarily).  This situation implies that
another test was needed.

This test used the "-.1" input string to specifically test this case,
which initially failed until the removed code was put back in.

Upon looking at the result code in the start and negative state process
functions and seeing the duplicate statements, a new valid matissa
character protected function was added to the base state class.