const-num: implement negate operator access function
TEST: verify that the negate operator access function returns false when
no negate operator is in the input stream.  The "-1-" input string with
a negative sign and subtract operator, but not a negate operator is used
for testing.  The following character is also checked for the subtract
operator.

To make the negate operator function false test compile, this function
needed to be implemented.  To make the test pass, it simply needed to
return false.

OTHER: rephrased the exit test case description removing the
"non-constant" word since the other exit condition to be implemented
(possible 'E' operator follows constant) will return a valid constant.