const-num: implement possible operator flag
TEST: verify when a possible operator is found ('E' followed by another
letter), the possible operator access function returns true. Also
verifies the number found and the next character after the 'E'.

A new possible operator flag was added, initialized to false in the
constructor, set to true in the set possible operator function, and
returned by the possible operator access function.