num-expr: implemented bare minimum class and parse function
TEST: compile a base numeric expression parser class and unit tests.

A new numeric expression unit test source file was added, with a single
test that instantiates a numeric expression parser, and calls its parse
function, but with no assertions.  The goal was to get this to compile
and run.

To accomplish this, a new numeric class was added.  The constructor
required an reference to an input stream, and the parse function takes
a code line and program references and returns a data type.  This
mirrors the constant number parser class.  The initial constructor
initialized the input stream reference member, and the parse function
simply returns a null data type.