num-expr: started pimpl pattern for numeric expression parser
TEST: verify compilation of pimpl pattern along with a working parse
function within the pimpl class, the verify the data type returning for
an integer constant in the input stream.

The pointer to implementation pattern allow the guts of a class to be
hidden from users of the main class' header file (which should speed
compilation of the header file).  The main class a single pointer to
the implementation instance with a forward declaration of the
implementation class.  A standard unique pointer is used for the pointer
so that it is freed automatically.  In order to use this in the header,
the main class 'must' have a desctructor declared and it must be defined
in the source file (an in line function would required the entire
implementation class definition to be visible).

The program code line and program unit arguments was moved from the
parse function to the constructor so that they can be passed to the
implementation constructor.  This also saves having to pass these
arguments all through the implementation class member functions.  A
starter parse operand function was added that just returns the integer
data type (to make the test pass).  The main parse function calls this
function.  (Eventually the other various parse operator functions will
be added.)