num-expr: implemented the parsing expressions with a single constant
TEST: verify an integer constant is parsed into the program line code.

The test was changed to use the require integer operand previously
defined for the constant number unit tests (the macro was put into a
new support header file so that it can be shared).

The test was made to pass be adding the creation of a constant number
parser instance in the parse operand function.  The parse function is
called and the data type returned.

The test continued to failed, which was caused by the implementation
class not defining the program code line member as a reference (so its
modified contents was only local to the class).  Also made the program
unit member variable a reference (don't want the program being copied).

OTHER: corrected the description of the test.