num-expr: add message and column to error when no more in input stream
TEST: skip over all characters in input stream before the last, verify
the message and column of the error throw.

The test was a bit tricky to get working right.  Instead of testing an
empty input stream, a dummy word plus a space was put in the stream.
The word is skipped over (by reading it), and then the space is read.
The std::ws function could not be used because it reads to the end
causing the input stream to be invalid.

In order for the correct column to be returned, the column must be
obtained before calling the constant number parser, otherwise when the
EOF is read from the input stream it will no longer be valid (will be
-1).  Therefore, the parse operand function was modified to save the
column first, and use that column if an error is to be throw.  The throw
was also given the appropriate error message.

Note: This change will also report errors for a negate operator
detected, since the negate operator has not been implemented yet.