num-expr: refactored class into a function operator class
For a class that has a single function, a function operator class
pattern is appropriate.  Therefore, the numeric expression parser class
was changed to this pattern (and the header file was cleaned up a bit).
The unit tests were updated accordingly, and the results are a little
clearer.

A small change was made to the private implementation class where a
parse expression function was added, which the main class function
operator function now calls.  The reason for this change is that this
function will need to be called internally by implementation class, for
example, when parsing a function, the parse expression function will be
called for the numeric arguments of the function.  Alternatively, a new
numeric expression parser instance could be created, but this would be
overkill (though will be needed for a string argument to get a string
expression).  Also, it self documents what the main's function operator
function does (parses an expression).