num-expr: added expected data type argument
TEST: verify code compiles and passes test.

An expected data type needs to be passed to the expression parser, so
that the caller can request a specific type of expression.  For example,
the SIN() function needs a double expression, but if the expression is
an integer expression, an internal CVTDBL code needs to be inserted.
However, a PRINT command can handle any type of expression (the exact
print code needed will be determine by the type of the expression),
therefore the PRINT command does not need to request a specific type, so
the null data type will be used to tell the expression parser that any
type is acceptable and that no convert codes are needed.

So to this end, an expected data type argument was added to the main
function operator function and to the implementation parse expression
function.  However, this argument is not currently used since the first
caller, the PRINT command will be passing a null data type, no code
needs to be implement right now to worry about conversions.