const-num: changed to pimpl and function operator pattern
The same PIMPL and function operator patterns changed made to the
expression parser class was made to the constant number parser class.
There were quite a few access functions in the constant number parser
class that are used by the state classes, but other users of this class
done need.  This was the reason the implementation forward declaration
needed to be public instead of private within the class.

OTHER: corrected some inconsistencies with the expression parser code
including the name of the include blocker (shouldn't begin with '_ as
these names are resevred), and pointer to the implementation instance
did not start with a 'p' for pointer.