commands: added initial command parser class and unit tests:
TEST: verify the code compiles, is able to take a blank input string
(simulating a blank line) and doesn't encode anything into to the
program code.

The same PIMPL, function operator, and arguments used for the constant
number and expression parsers was used for the initial command parser
and its implementation classes.  Through not strictly necessary to make
the test pass, the function operator function calls the parse function
of the implementation class.

The make the test pass, the parse function doesn't need to do anything
and the program code will be left empty.  As more code is implemented,
this test will need to continue to pass.